test(vue-query): fixed test for useQueries with skipToken
